Fix code logic and modify functions names
Modify principal functions names to more
appropriate names, and fix the logic of the code
to consider user inputs not working in the last
commit.

Fix bugs when clicking "=" when no operation has been
initiated that caused errors in calculations.
Fix bugs when wrong input has been made by user
(example division by 0).
Add a new roundFloats function that calculate number
of decimals of precisions it is needed to be taken
in count when calculating and displaying results.